Variant,VariantDict: More type lookup methods
In ostree/rpm-ostree we use GVariant extensively and particularly
`a{sv}`s.  I have this helper trait
https://github.com/ostreedev/ostree-rs-ext/blob/6fcdde1697a851f9c29d9f8b15d77e114a7a5ae7/lib/src/variant_utils.rs#L32
but we can do better by offering an API equivalent to
`lookup()` + `Variant::get()` that returns the native type.

Plus for `Variant` it's often convenient to have a version that
returns a `Result<T>` with a useful error message about
the actual+expected types.