glib/variant: Expose `g_variant_get_normal_form()`
In ostree we compute `sha256(variant)` and expect this to be
in "normal form".  While this is a relatively obscure feature
it is public GLib API and trivial to expose.

The main benefit is this will help us drop a dependency on
`glib-sys` and move towards the re-export model from the 0.14
series.