variant: Add more APIs to access typed children
A previous PR added `try_child_value()`; this adds a corresponding
`try_child_get()` which I think more users are going to want.

While we're here add `child_get()` which panics on any
unmet constraints.  In many cases such as e.g. iteration one
would be able to statically know that the bounds and types
are correct.

And another reason to add these is that as the `TODO` says,
in the future we can optimize these to avoid heap allocations
for intermediate `GVariant` references.