variant: Add an API to iterate over `as`
In ostree and rpm-ostree in particular we have quite a lot of
variants of type `as` (array of string).  Right now, accessing
those requires lots and lots of allocations.  Honestly, it's
mostly not truly performance sensitive, but I'd like to avoid
doing obviously silly things so that it doesn't become
a problem down the line.

We've had `.str()` for quite a while to access a variant of type `s`
without copying out the string data.  This extends that model to
variants of type `as` by allowing iteration over the values.