glib: Merge variants of `wrapper!` as much as possible
Handling object wrappers with or without `@extends` and with or without
`@implements` in one variant is problematic, since a comma between the
two is necessary if and only if both are used.

But assuming no issues turn up, this should be a good improvement,
making this a fair bit simpler.