glib: Protect more against main context channel GSources and non-Send future sources are dropped from the wrong thread
This can only really happen if the application somehow retrieves the
GSource via the main context API and sends it to another thread, but we
can often do better than directly panicking by sending the non-Send data
to the main context if it still exists.

Otherwise there's no way around panicking other than leaking, and
panicking seems like a better approach here as it would be an
application bug if this ever happens.