glib: Relax checks for functions spawning non-Send closures/futures on a main context
Only try to acquire the main context and panic if that fails, and not
already if the current thread is not yet the owner of the main context.

This allows these operations to succeed before the main context was
explicitly acquired directly or via `push_thread_default()`.

As part of this also guard a bit more against main contexts changing
ownership from one thread to another. If this ever happened before, a
non-Send closure would've been called from a different thread. Now this
simply panics as it should.