glib: Fix bug in main context channel test that caused it to deadlock
With edition 2021 the closures are only borrowing what they have to, so
here in this case it didn't actually borrow the helper as it was not
actually used inside the closure. With that, it was not dropped when the
main context channel was dropped and the main loop was never quit.