glib: Fix `FromValue` impls for `&Boxed`, `&Shared` and `&Object`
Instead of working with a pointer to the value's pointer that is
allocating temporarily on the stack, work with the pointer to the
memory location where the value is stored inside the `GValue`.

The first one is only valid inside the scope of `from_value`, the latter
is valid as long as the `Value` is valid, which is what we actually
want.

This happened to work before in the tests but was technically incorrect,
would've fallen apart if something else re-used the same stack space
later and valgrind was not very happy about it.