gio: add type constraints for some task methods
In particular require ToSendValue for using run_in_thread.
run_in_thread now takes a closure that returns a result with
the value, this allows the compiler to infer the type of
the generic and also makes the closure more ergonomic.

return_value and propagate_value are also changed to be
generic and perform conversion to and from value implicitely.