Reimplement gio::Task manually
Use the newly added ability to have generic types in the wrapper
macro and make Task generic on the ValueType it will return.

This fixes some soundness holes in the Task api:
1. The task itself is Send only if the value type is Send, thus preventing
   the caller to return the task value in a thread, pass the task to
   another thread, and then propagating the value, if the value is not
   Send
2. The Task::run_in_thread method is only implemented for Send values
3. Task::run_in_thread requires source object to be Send

Additionally the task api is more ergonomic as you now return and
propagate the actual type rather than a glib::Value.