gio: further rework Task API by splitting LocalTask and Task
This fixes further soundness issues in the gio task api.
We need to pass callbacks that require Send when using tasks with threads.

To do this we need to split the API into two types, LocalTask (which does
not require Send) and Task, which does require Send and which implements
run_in_thread. Since the implementation is almost identical, it is
generated through a macro.

There is a further soundness issue with the return_result and propagate
functions, since these can only be called once. With the API of GTask
we do not have a way to enforce this, so instead they are marked as
unsafe for the threaded case.