gio: mark creating Task as unsafe and add public documentation
Even with all the reccent improvements to the Task API, it is
still possible to bypass the requirements, for example by
casting LocalTask to Task, since they have the same underlying
GType.

At this point mark the constructors as unsafe and document
which invariants need to be manually esured by the callers.

The goal of this API is to enable callers to implement in
Rust APIs exposed to the C ffi which follow the glib
convention for async calls. Pure Rust applications should
probably use better abstractions to dispatch asynchronous
tasks.