glib: Add an `Object` specific value type checker
This doesn't only check the type of the `Value` itself but if there's a
mismatch it grabs the contained object and does a check against the
concrete type of the object too.

This is necessary if e.g. the `Value` is initialized for a plain
`Object`, a `gtk::Widget` is stored inside it and it is tried to be
retrieved as `gtk::Widget` instead of `Object`.