glib: Don't use `from_glib_full()` for the individual items of an array for boxed inline types
There is only a single allocation for the whole array and the individual
items should not be freed but just moved out of the array into the
`Vec`.

Only in the case of an array of pointers to boxed inline types
`from_glib_full()` should be used to free the individual allocations.

Fixes https://github.com/gtk-rs/gtk4-rs/issues/932