glib: Only auto-impl `Send+Sync` for subclasses if the parent type does
Previously a `Send+Sync` subclass of a `gtk::Widget` would automatically
be `Send+Sync` too, which is obviously not allowed.

To avoid this, also add the parent type as `PhantomData` to the wrapper
struct so that this is considered too.

Handle `glib::Object` specially here because while it is not `Send+Sync`
itself, subclasses of it can be `Send+Sync` just fine. The only reason
why `glib::Object` is not is because there are some subclasses that are
not and otherwise this could be circumvented by upcasting.