gio: Manually implement ActionEntry
It is pretty useful for ActionMap implementors like GtkApplication
and GtkWindow. It avoids usually the usage of the clone! macro
as we pass the ActionMap to the activate/change_state callbacks.

Especially that Gtk doesn't provide a
gtk_application_class_install_action
alternative to the gtk_widget_class_install_action.