glib/types: use rustc-hash for HashMap
While profiling the standalone threadshare benchmark in
gst-plugins-rs, callgrind showed 2% cycles spent in
`ElementImplExt::panicked`. This function retrieves the
panicking status in trampolines and makes use of
`instance_data` from `ObjectSubclass`, which is stored in
a `HashMap`.

Rust's `std::collections::HashMap` is known to make use of
a "hashing algorithm selected to provide resistance against
HashDoS attacks" [1]. Since the `intance_data` is indexed
by a `glib::Type`, I thought it would make sense to test a
less strict but quicker hasher.

I first tested `fnv` [2] which is an implementation of the
Fowler-Noll-Vo hash function which is used in major crates
such as `hashbrown`. The crate `twox-hash` [3] features a
benchmark showing larger throughputs for `fnv` with small
data. Callgrind reports a 50% reduction in cycle estimation
for `ElementImplExt::panicked`. I also observe a slight
reduction in CPU usage with `ts-standalone`.

`rustc-hash` [4] is used by `rustc` where it "consistently
out-performs adn FNV-based hash 5 [...] -- the collision
rate is similar or slightly worse then FNV, but the speed
of the has function itself is much higher". Callgrind
reports 70% reduction in cycle estimation for `panicked`
compared to the default hasher. CPU usage and stats seem
similar to those obtained with `fnv` for `ts-standalone`.

`rustc-hash` was adopted.

[1]: https://doc.rust-lang.org/std/collections/struct.HashMap.html
[2]: https://crates.io/crates/fnv
[3]: https://crates.io/crates/twox-hash
[4]: https://crates.io/crates/rustc-hash