cairo-rs: add Context API to enforce stack correctness
`Context` has the `Context::save` and `Context::restore` methods to
save the current Context state, do changes to the new state, and be able
to restore the original state. If a save() is not followed by a
restore() the result is visual bugs (at least in my experience).

In C there's no way to prevent a user from forgetting a restore().

In Rust you can nest stack levels naturally using the borrowing rules.
Since a mutable reference of a variable prevents the variable to be used
while the reference is still alive, we can ensure the user cannot create
a new stack level and forget about the previous one. The restore() is
performed inside the Drop implementation. That means that if it
compiles, the stacking is correct, and since this is on the type level
it is a zero-cost abstraction.

Questions:
- should `Context` be available to the user without `unsafe`? As long as
  it is, the user can create as many `ContextRef`s as they wish and bypass
  this API.
- If restore().unwrap() in the Drop impl panics, the process aborts. Is
  that acceptable?