glib/GStringPtr: Add `as_str()` and `Deref<Target=&str>`
I was updating some code to glib 0.18 and we were previously
using `as_str()` - which is what the default Rust nomenclature
for an *infallible* reference-to-reference conversion uses.

The `to_str()` is used for fallible conversions
(e.g. https://doc.rust-lang.org/std/ffi/struct.OsStr.html#method.to_str)
which this is not.