glib: Fix `MatchInfo::next()` handling of returning `FALSE`
`FALSE` is returned both on errors when the `GError*` is set as well as
if there is no further match, in which case the `GError*` is not set.

Change the return type from `Result<(), glib::Error>` to
`Result<bool, glib::Error>` to be able to represent that and don't
assert that either `TRUE` is returned or the `GError*` is set.

Fixes https://github.com/gtk-rs/gtk4-rs/issues/1749