glib-macros: Don't produce unnecessary braces in `clone!(async move { x })`
This would wrongly suggest to remove the braces, which is invalid
syntax, because the macro already adds another layer of braces around
the body of the async block.

To avoid this, don't output the braces of the async block in the macro
expansion and instead only use the braces added by the macro.