glib: Add support for registering GTypes with name conflicts
When explicitly allowed when defining the type, a GType name conflict
wouldn't cause a panic anymore but instead a new name is found by
appending a counter to the given name.

This is useful if multiple definitions of the same type can end up in
the same process.

Also use this for `glib::BoxedAnyObject`, `gio::WriteOutputStream` and
`gio::ReadInputStream`.

Fixes https://github.com/gtk-rs/gtk-rs-core/issues/1430