glib: Collation keys are not necessary valid UTF-8
All that is guaranteed about them is that they give the correct order
for the original strings when passed to strcmp().

Also add a separate integration test for this that calls setlocale() to
get the default system locale. This makes sure that not just the "C"
locale is selected, which returns literally the same string as collation
key.

Fixes https://github.com/gtk-rs/gtk-rs-core/issues/1504