glib-macros: Properties: allow setting getter and setter docs
In the previous commit we made it so we copy the docs for a `#[property]` in to
the getter. This was an improvement but it still did not allow for any doc comments
on the setter that might be relevant only there.

Here we introduce the ability to specify doc comments for the getter and setter
individually via the use of the header syntax to switch between them. By writing
`# Getter` or `# Setter` in a line, you make the lines following it go to the
getter or setter function definitions.

If neither of these values exist, we copy the doc comments to both. If you use
these values, then anything in the doc comments before the first one is
discarded.