gio: Rework `glib::ExitCode` for `gio::Application`
Add a manual override to use `glib::ExitCode` or `ControlFlow<glib::ExitCode>`
instead of the gir generated `i32` return value for the `command-line`
and `handle-local-options` signals as well as the `local_command_line`
virtual function.

This reworks the `ExitCode` type to only support `u8` values. The `From<i32>`
implementation is replaced with a `TryFrom<i32>` implementation.