gio: add methods for returning and propagating boolean and int in/from a `Task`
When using Rust we prefer to use a `Value` for everything but gio also lets you
return explicitly a `gboolean` or `gssize` which avoid boxing.

Usage of these functions to return these kinds of values requires a call to the
corresponding propagate function and you cannot mix them. Trying to extract a
`Value` when either of them were used leads to an incorrect value.

Expose methods in `Task` and `LocalTask` to return and propagate these kinds
instead of assuming everything can go through a `Value`.