gio: document the interaction of the `Task` result setter and getters
The three ways you can return a result from the task must be match in how you
set and get the fields.

Document this in the different functions and mark them as unsafe unconditionally
to indicate that there is no way within the type system to ensure that you are
using it safely.