Add subscribe_to_signal
Add a new method subscribe_to_signal which returns a managed signal
subscription which unsubscribes from the signal when dropped.  Also add
a weak ref counterpart to this signal subscription to allow breaking
reference cycles.

This fits better into Rust's ownership model and makes it much easier to
handle the lifetime of a signal subscription in async code.

Mark the old signal_subscribe and signal_unsubscribe as deprecated in
favour of this new subscribe_to_signal method.