glib: Add Error::with_domain constructor and code getter
Add a constructor to create errors from a domain quark and code,
useful when creating errors with the same domain/code but different
messages as we can't get a rust type which implements the `ErrorDomain`
in that case.

This also adds a code() getter method to retrieve the error code.